Enrich AppendEntries return values and fill in business logic to optimize the number of them
The goal is to exchange less AppendEntries RPCs with followers that are far behind and not
only contain some (or none) of the conflicting entries, but also some of the latter entries
in their log are not in synch with the leader, so they need to be overwritted.

The optimization allows the leader to jump whole terms (whatever number of indexes they
contain) in the log, whenever they detect that they do not hold the specific term on their
log.

In another way, when the optimization was non-existent, the leader neeeded to send as many
AppendEntries RPCs as the number of entries that belong to the conflicting term.

Due to the relatively small and infrequent number of failures in LogCabin servers, this optimization
is thought to be “overkill”.